requirements.txt
# this file isn't yet reliable!!
google-cloud-storage

boto3==1.26.75
botocore==1.29.75
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
jmespath==1.0.1
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14